move stories components to own folder
