+++ components/stories/StoriesAvatar.tsx
