+++ components/stories/StoriesCoverImage.tsx
